backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
blinker==1.8.2
    # via flask
click==8.1.7
    # via flask
flask==3.0.3
    # via
    #   -r requirements.in
    #   flask-sqlalchemy
flask-sqlalchemy==3.1.1
    # via -r requirements.in
greenlet==3.0.3
    # via sqlalchemy
iniconfig==2.0.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
packaging==24.1
    # via pytest
pluggy==1.5.0
    # via pytest
ply==3.11
    # via -r requirements.in
psycopg2==2.9.9
    # via -r requirements.in
pytest==8.2.2
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
sqlalchemy==2.0.30
    # via
    #   -r requirements.in
    #   flask-sqlalchemy
typing-extensions==4.12.2
    # via sqlalchemy
werkzeug==3.0.3
    # via flask
flask-cors==4.0.1
